requirement.txt
APScheduler==3.10.4
asgiref==3.8.0
async-timeout==4.0.3
bcrypt==4.1.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.1.13
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
google-auth==2.29.0
idna==3.6
mysqlclient==2.2.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pymongo==3.12.3
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
